# docs/swagger.yaml
openapi: 3.0.3
info:
    title: UnaHur Anti-Social Net API
    description: |
        API REST del backend para la red social **UnaHur Anti-Social Net**.
        Permite gestionar usuarios, publicaciones, comentarios, etiquetas, imágenes asociadas a posts
        y relaciones de seguimiento entre usuarios (followers).

        **Variables de entorno relevantes**

        | Variable | Descripción |
        | --- | --- |
        | PORT | Puerto del servidor (por defecto 3001) |
        | MESES | Antigüedad máxima en meses para comentarios en publicaciones (por defecto 6) |
        | CACHE_POSTS_ENABLED | Caché en memoria para GET de posts (por defecto true) |
        | CACHE_POSTS_TTL_SECONDS | TTL del caché en segundos (por defecto 60) |
        | DB_DIALECT | sqlite, mysql, mariadb o postgres |
        | ENABLE_SWAGGER | Habilita Swagger UI en producción (por defecto false) |
    version: 1.0.0
    contact:
        name: Anti-Social Team
servers:
    - url: http://localhost:3001
      description: Servidor local (puerto configurable vía variable de entorno PORT)
...
paths:
...
    /posts:
        get:
            tags: [Posts]
            summary: Listar publicaciones
            description: |
                Respuesta cacheada en memoria (TTL configurable). Se invalida al crear, editar o borrar posts o comentarios.
            operationId: getAllPosts
            parameters:
                - name: user_id
                  in: query
                  description: Filtrar publicaciones por ID de usuario
                  required: false
                  schema:
                      type: integer
                      minimum: 1
                - $ref: "#/components/parameters/ViewerIdQuery"
            responses:
                "200":
                    description: |
                        Lista de publicaciones con usuario, imágenes, tags y comentarios visibles.
                        Los comentarios con antigüedad mayor o igual a `MESES` no se incluyen (filtro por `monthsOld`).
                        Incluye posteos de usuarios con perfil público o privado. Si se filtra por `user_id`,
                        un perfil privado solo es accesible para el propio usuario o sus seguidores (`viewer_id`).
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/PostDetail"
                "400":
                    $ref: "#/components/responses/ValidationError"
                "403":
                    $ref: "#/components/responses/Forbidden"
                "404":
                    $ref: "#/components/responses/NotFound"
        post:
            tags: [Posts]
            summary: Crear una publicación
            operationId: createPost
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/PostCreate"
            responses:
                "201":
                    description: Publicación creada
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/PostDetail"
                "400":
                    $ref: "#/components/responses/ValidationError"
                "404":
                    $ref: "#/components/responses/NotFound"
    /posts/{id}:
        parameters:
            - $ref: "#/components/parameters/IdParam"
        get:
            tags: [Posts]
            summary: Obtener una publicación por ID
            description: |
                Devuelve la publicación con sus relaciones. Los comentarios más antiguos
                que X meses (configurable con la variable de entorno MESES) no se incluyen.
            operationId: getPostById
            responses:
                "200":
                    description: Publicación encontrada
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/PostDetail"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "404":
                    $ref: "#/components/responses/NotFound"
        patch:
            tags: [Posts]
            summary: Actualizar una publicación
            description: |
                Permite actualizar la descripción y/o reemplazar por completo la lista de etiquetas (`tags`).
                Si se envía `tags`, las etiquetas anteriores del post se sustituyen por las nuevas.
            operationId: updatePost
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/PostUpdate"
            responses:
                "200":
                    description: Publicación actualizada
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/PostDetail"
                "400":
                    $ref: "#/components/responses/ValidationError"
                "404":
                    $ref: "#/components/responses/NotFound"
        delete:
            tags: [Posts]
            summary: Eliminar una publicación
            operationId: deletePost
            responses:
                "200":
                    description: Publicación eliminada
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/MessageResponse"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "404":
                    $ref: "#/components/responses/NotFound"
    /posts/{id}/images:
        parameters:
            - $ref: "#/components/parameters/IdParam"
        get:
            tags: [Post Images]
            summary: Listar imágenes de una publicación
            operationId: getPostImagesByPost
            responses:
                "200":
                    description: Imágenes de la publicación
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/PostImage"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "404":
                    $ref: "#/components/responses/NotFound"
        post:
            tags: [Post Images]
            summary: Subir una imagen a una publicación
            operationId: createPostImageOnPost
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            required: [image]
                            properties:
                                image:
                                    type: string
                                    format: binary
                                    description: Archivo de imagen (JPEG, PNG o WebP, máx. 5 MB)
            responses:
                "201":
                    description: Imagen asociada al post
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/PostImage"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "404":
                    $ref: "#/components/responses/NotFound"
...
    /post-images:
        get:
            tags: [Post Images]
            summary: Listar todas las imágenes de posts
            operationId: getAllPostImages
            responses:
                "200":
                    description: Lista de imágenes
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/PostImage"
        post:
            tags: [Post Images]
            summary: Crear una imagen asociada a un post
            operationId: createPostImage
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            required: [image, postId]
                            properties:
                                postId:
                                    type: integer
                                    minimum: 1
                                    description: ID de la publicación
                                image:
                                    type: string
                                    format: binary
                                    description: Archivo de imagen (JPEG, PNG o WebP, máx. 5 MB)
            responses:
                "201":
                    description: Imagen creada
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/PostImage"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "404":
                    $ref: "#/components/responses/NotFound"
    /post-images/{id}:
        parameters:
            - $ref: "#/components/parameters/IdParam"
        get:
            tags: [Post Images]
            summary: Obtener una imagen por ID
            operationId: getPostImageById
            responses:
                "200":
                    description: Imagen encontrada
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/PostImage"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "404":
                    $ref: "#/components/responses/NotFound"
        patch:
            tags: [Post Images]
            summary: Actualizar una imagen
            operationId: updatePostImage
            requestBody:
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                postId:
                                    type: integer
                                    minimum: 1
                                    description: Reasignar imagen a otro post (opcional)
                                image:
                                    type: string
                                    format: binary
                                    description: Nueva imagen (opcional)
            responses:
                "200":
                    description: Imagen actualizada
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/PostImage"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "404":
                    $ref: "#/components/responses/NotFound"
        delete:
            tags: [Post Images]
            summary: Eliminar una imagen
            operationId: deletePostImage
            responses:
                "200":
                    description: Imagen eliminada
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/MessageResponse"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "404":
                    $ref: "#/components/responses/NotFound"
    /comments:
        get:
            tags: [Comments]
            summary: Listar todos los comentarios
            description: |
                Devuelve comentarios con usuario y publicación asociados (sin filtro por antigüedad).
                Opcionalmente filtra por `post_id`.
            operationId: getAllComments
            parameters:
                - name: post_id
                  in: query
                  description: Filtrar comentarios por ID de publicación
                  required: false
                  schema:
                      type: integer
                      minimum: 1
            responses:
                "200":
                    description: Lista de comentarios con usuario y post
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/CommentDetail"
                "400":
                    $ref: "#/components/responses/ValidationError"
                "404":
                    $ref: "#/components/responses/NotFound"
        post:
            tags: [Comments]
            summary: Crear un comentario
            operationId: createComment
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/CommentCreate"
            responses:
                "201":
                    description: Comentario creado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/CommentDetail"
                "400":
                    $ref: "#/components/responses/ValidationError"
                "404":
                    $ref: "#/components/responses/NotFound"
    /comments/{id}:
        parameters:
            - $ref: "#/components/parameters/IdParam"
        get:
            tags: [Comments]
            summary: Obtener un comentario por ID
            operationId: getCommentById
            responses:
                "200":
                    description: Comentario encontrado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/CommentDetail"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "404":
                    $ref: "#/components/responses/NotFound"
        patch:
            tags: [Comments]
            summary: Actualizar un comentario
            operationId: updateComment
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/CommentUpdate"
            responses:
                "200":
                    description: Comentario actualizado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/CommentDetail"
                "400":
                    $ref: "#/components/responses/ValidationError"
                "404":
                    $ref: "#/components/responses/NotFound"
        delete:
            tags: [Comments]
            summary: Eliminar un comentario
            operationId: deleteComment
            responses:
                "200":
                    description: Comentario eliminado
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/MessageResponse"
                "400":
                    $ref: "#/components/responses/BadRequest"
                "404":
                    $ref: "#/components/responses/NotFound"
...
components:
    parameters:
        IdParam:
            name: id
            in: path
            required: true
            description: Identificador numérico del recurso
            schema:
                type: integer
                minimum: 1
...
        ViewerIdQuery:
            name: viewer_id
            in: query
            required: false
            description: ID del usuario que consulta el perfil (para validar perfiles privados)
            schema:
                type: integer
                minimum: 1
    responses:
        BadRequest:
            description: Solicitud inválida
            content:
                application/json:
                    schema:
                        oneOf:
                            - $ref: "#/components/schemas/MessageResponse"
                            - $ref: "#/components/schemas/ValidationErrorResponse"
        ValidationError:
            description: Error de validación de datos
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ValidationErrorResponse"
        NotFound:
            description: Recurso no encontrado
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/MessageResponse"
        Forbidden:
            description: Acceso denegado al perfil privado
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/MessageResponse"
...
    schemas:
        MessageResponse:
            type: object
            properties:
                message:
                    type: string
                    example: Operación realizada correctamente
...
        ValidationErrorResponse:
            type: object
            properties:
                errores:
                    type: array
                    items:
                        type: object
                        properties:
                            atributo:
                                type: string
                            error:
                                type: string
...
        PostCreate:
            type: object
            required: [titulo, description, user_id]
            properties:
                titulo:
                    type: string
                    minLength: 1
                    maxLength: 200
                description:
                    type: string
                    minLength: 1
                    maxLength: 5000
                user_id:
                    type: integer
                    minimum: 1
                tags:
                    type: array
                    minItems: 1
                    items:
                        type: string
                        minLength: 1
                        maxLength: 50
        PostUpdate:
            type: object
            minProperties: 1
            description: Al menos uno de `titulo`, `description` o `tags` es requerido.
            properties:
                titulo:
                    type: string
                    minLength: 1
                    maxLength: 200
                description:
                    type: string
                    minLength: 1
                    maxLength: 5000
                tags:
                    type: array
                    minItems: 1
                    description: Reemplaza todas las etiquetas actuales del post.
                    items:
                        type: string
                        minLength: 1
                        maxLength: 50
        PostDetail:
            type: object
            properties:
                id:
                    type: integer
                titulo:
                    type: string
                description:
                    type: string
                user_id:
                    type: integer
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
                user:
                    $ref: "#/components/schemas/UserSummary"
                postImages:
                    type: array
                    items:
                        $ref: "#/components/schemas/PostImage"
                tags:
                    type: array
                    items:
                        $ref: "#/components/schemas/TagSummary"
                comments:
                    type: array
                    items:
                        $ref: "#/components/schemas/CommentSummary"
        UserSummary:
            type: object
            properties:
                id:
                    type: integer
                nickname:
                    type: string
                name:
                    type: string
                lastName:
                    type: string
                profilePicture:
                    type: string
                    format: uri
                    nullable: true
        PostImage:
            type: object
            properties:
                id:
                    type: integer
                url:
                    type: string
                    format: uri
                    example: http://localhost:3001/uploads/posts/imagen.jpg
                post_id:
                    type: integer
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
        CommentCreate:
            type: object
            required: [content, post_id, user_id]
            properties:
                content:
                    type: string
                    minLength: 1
                    maxLength: 150
                post_id:
                    type: integer
                    minimum: 1
                user_id:
                    type: integer
                    minimum: 1
        CommentUpdate:
            type: object
            minProperties: 1
            properties:
                content:
                    type: string
                    minLength: 1
                    maxLength: 150
        CommentSummary:
            type: object
            properties:
                id:
                    type: integer
                content:
                    type: string
                monthsOld:
                    type: integer
                    description: Meses transcurridos desde la creación (campo virtual)
        PostSummary:
            type: object
            properties:
                id:
                    type: integer
                description:
                    type: string
        TagSummary:
            type: object
            properties:
                id:
                    type: integer
                name:
                    type: string
        CommentDetail:
            type: object
            properties:
                id:
                    type: integer
                content:
                    type: string
                user_id:
                    type: integer
                post_id:
                    type: integer
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
                monthsOld:
                    type: integer
                    description: Meses transcurridos desde la creación (campo virtual). En GET de posts, si es >= MESES el comentario no se devuelve.
                user:
                    $ref: "#/components/schemas/UserSummary"
                post:
                    $ref: "#/components/schemas/PostSummary"
...